requirements.txt
# Django and Core Dependencies
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.6.0
django-filter==24.3
channels==4.1.0
channels-redis==4.2.0

# Database
psycopg[binary]==3.2.3
dj-database-url==2.3.0
redis==5.2.0
django-redis==5.4.0

# Celery for Background Tasks
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1

# Monitoring and Metrics
pysnmp==6.2.6
prometheus-client==0.21.1
django-prometheus==2.3.1

# Authentication and Security
django-otp==1.5.4
cryptography==42.0.8

# Data Processing
openpyxl==3.1.5
pandas==2.2.3
numpy==2.1.3

# Network Operations
paramiko==3.5.0
requests==2.32.3

# Machine Learning (for anomaly detection)
scikit-learn==1.5.2

# Development and Testing
pytest==8.3.3
pytest-django==4.9.0
hypothesis==6.112.2
factory-boy==3.3.1
coverage==7.6.7

# Production
gunicorn==23.0.0
whitenoise==6.8.2

# Utilities
python-decouple==3.8
Pillow==10.4.0